refactor: 사용하지 않는 클래스 삭제 및 Order 패키지 이동
- OrderChecker와 OrderProcessor 클래스 삭제
- Order를 domain.order -> domain 패키지로 이동
