+++ src/main/java/store/service/OrderService.java
@@ -3,7 +3,7 @@ package store.service;
 import java.util.ArrayList;
 import java.util.List;
 import store.controller.InputHandler;
-import store.domain.order.Order;
+import store.domain.Order;
 import store.dto.OrderCheckDto;
 import store.dto.OrderResultDto;
